requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
djoser==2.2.0
django-imagekit==4.1.0